Add submission graphs to discussion details in teacher
refs: MBL-14386
affects: Teacher
release note: none
