+++ Core/Core/Discussions/DiscussionDetailsViewController.swift
@@ -20,6 +20,7 @@ import UIKit
 import WebKit
 
 public class DiscussionDetailsViewController: UIViewController, ColoredNavViewProtocol, ErrorViewController {
+    @IBOutlet weak var dueSection: UIView!
     lazy var optionsButton = UIBarButtonItem(image: .icon(.more), style: .plain, target: self, action: #selector(showTopicOptions))
     @IBOutlet weak var pointsLabel: UILabel!
     @IBOutlet weak var pointsView: UIView!
@@ -29,6 +30,8 @@ public class DiscussionDetailsViewController: UIViewController, ColoredNavViewPr
     let refreshControl = CircleRefreshControl()
     @IBOutlet weak var scrollView: UIScrollView!
     @IBOutlet weak var spinnerView: CircleProgressView!
+    @IBOutlet public weak var sectionsStack: UIStackView!
+    @IBOutlet weak var submissionsSection: UIView!
     public var titleSubtitleView = TitleSubtitleView.create()
     @IBOutlet weak var titleLabel: UILabel!
     @IBOutlet weak var webViewPlaceholder: UIView!
@@ -215,18 +218,30 @@ public class DiscussionDetailsViewController: UIViewController, ColoredNavViewPr
         pointsLabel.text = assignment?.first?.pointsPossibleText
         pointsView.isHidden = assignment?.first?.pointsPossible == nil || showRepliesToEntryID != nil
 
-        if topic.first?.published == true {
-            publishedIcon.image = .icon(.publish, .solid)
-            publishedIcon.tintColor = .named(.textSuccess)
-            publishedLabel.text = NSLocalizedString("Published", bundle: .core, comment: "")
-            publishedLabel.textColor = .named(.textSuccess)
-        } else {
-            publishedIcon.image = .icon(.no, .solid)
-            publishedIcon.tintColor = .named(.textDark)
-            publishedLabel.text = NSLocalizedString("Unpublished", bundle: .core, comment: "")
-            publishedLabel.textColor = .named(.textDark)
-        }
+        let isPublished = topic.first?.published == true
+        publishedIcon.image = .icon(isPublished ? .publish : .no, .solid)
+        publishedIcon.tintColor = .named(isPublished ? .textSuccess : .textDark)
+        publishedLabel.text = isPublished
+            ? NSLocalizedString("Published", bundle: .core, comment: "")
+            : NSLocalizedString("Unpublished", bundle: .core, comment: "")
+        publishedLabel.textColor = .named(isPublished ? .textSuccess : .textDark)
         publishedView.isHidden = env.app != .teacher || isAnnouncement || showRepliesToEntryID != nil
+    
+        if let assignmentID = topic.first?.assignmentID {
+            if dueSection.isHidden {
+                dueSection.isHidden = false
+            }
+
+            if topic.first?.groupTopicChildren == nil, submissionsSection.isHidden {
+                let controller = SubmissionBreakdownViewController.create(
+                    context: context,
+                    assignmentID: assignmentID,
+                    submissionTypes: [.discussion_topic]
+                )
+                embed(controller, in: submissionsSection)
+                submissionsSection.isHidden = false
+            }
+        }
 
         render()
     }
