+++ Core/Core/Submissions/GetSubmissionComments.swift
@@ -16,8 +16,8 @@
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 //
 
-import Foundation
 import CoreData
+import Foundation
 
 public class GetSubmissionComments: APIUseCase {
     let assignmentID: String
