--- Core/Core/Use Cases/GetSubmission.swift
@@ -1,62 +1,0 @@
-//
-// This file is part of Canvas.
-// Copyright (C) 2018-present  Instructure, Inc.
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU Affero General Public License as
-// published by the Free Software Foundation, either version 3 of the
-// License, or (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU Affero General Public License for more details.
-//
-// You should have received a copy of the GNU Affero General Public License
-// along with this program.  If not, see <https://www.gnu.org/licenses/>.
-//
-
-import Foundation
-import CoreData
-
-public class GetSubmission: APIUseCase {
-    public let context: Context
-    public let assignmentID: String
-    public let userID: String
-
-    public typealias Model = Submission
-
-    public init(context: Context, assignmentID: String, userID: String) {
-        self.assignmentID = assignmentID
-        self.userID = userID
-        self.context = context
-    }
-
-    public var cacheKey: String? {
-        return "get-\(context.id)-\(assignmentID)-\(userID)-submission"
-    }
-
-    public var request: GetSubmissionRequest {
-        return GetSubmissionRequest(context: context, assignmentID: assignmentID, userID: userID)
-    }
-
-    public var scope: Scope {
-        return Scope(
-            predicate: NSPredicate(
-                format: "%K == %@ AND %K == %@",
-                #keyPath(Submission.assignmentID),
-                assignmentID,
-                #keyPath(Submission.userID),
-                userID
-            ),
-            order: [NSSortDescriptor(key: #keyPath(Submission.attempt), ascending: false)]
-        )
-    }
-
-    public func write(response: APISubmission?, urlResponse: URLResponse?, to client: NSManagedObjectContext) {
-        guard let item = response else {
-            return
-        }
-        Submission.save(item, in: client)
-    }
-}
