+++ Core/CoreTests/Submissions/APISubmissionTests.swift
@@ -20,7 +20,7 @@ import Foundation
 import XCTest
 @testable import Core
 
-class APISubmissionRequestableTests: CoreTestCase {
+class APISubmissionTests: CoreTestCase {
     func testGetSubmissionRequest() {
         XCTAssertEqual(GetSubmissionRequest(context: ContextModel(.course, id: "1"), assignmentID: "2", userID: "3").path, "courses/1/assignments/2/submissions/3")
         XCTAssertEqual(GetSubmissionRequest(context: ContextModel(.course, id: "1"), assignmentID: "2", userID: "3").query, [ APIQueryItem.array("include", [
