+++ Core/CoreTests/Submissions/SubmissionTests.swift
@@ -17,6 +17,7 @@
 //
 
 import Foundation
+import MobileCoreServices
 import XCTest
 @testable import Core
 
@@ -261,5 +262,94 @@ class SubmissionTests: CoreTestCase {
 
         let notSubmitted = Submission.make(from: .make(submitted_at: nil, late: false, missing: false))
         XCTAssertEqual(notSubmitted.status, .notSubmitted)
+    }
+}
+
+class SubmissionTypeTests: XCTestCase {
+    func testInitRawValue() {
+        // Converting to & from String is needed by database models
+        XCTAssertEqual(SubmissionType(rawValue: "discussion_topic"), .discussion_topic)
+        XCTAssertEqual(SubmissionType.discussion_topic.rawValue, "discussion_topic")
+    }
+
+    func testLocalizedString() {
+        XCTAssertEqual(SubmissionType.discussion_topic.localizedString, "Discussion Comment")
+        XCTAssertEqual(SubmissionType.external_tool.localizedString, "External Tool")
+        XCTAssertEqual(SubmissionType.media_recording.localizedString, "Media Recording")
+        XCTAssertEqual(SubmissionType.none.localizedString, "No Submission")
+        XCTAssertEqual(SubmissionType.not_graded.localizedString, "Not Graded")
+        XCTAssertEqual(SubmissionType.online_quiz.localizedString, "Quiz")
+        XCTAssertEqual(SubmissionType.online_text_entry.localizedString, "Text Entry")
+        XCTAssertEqual(SubmissionType.online_upload.localizedString, "File Upload")
+        XCTAssertEqual(SubmissionType.online_url.localizedString, "Website URL")
+        XCTAssertEqual(SubmissionType.on_paper.localizedString, "On Paper")
+    }
+
+    func testAllowedMediaTypesForMediaRecordings() {
+        var submissionTypes: [SubmissionType] = [.media_recording]
+        XCTAssertEqual(submissionTypes.allowedMediaTypes, [kUTTypeMovie as String, kUTTypeAudio as String])
+
+        submissionTypes = [.media_recording, .online_upload]
+        XCTAssertEqual(submissionTypes.allowedMediaTypes, [kUTTypeMovie as String, kUTTypeImage as String])
+    }
+
+    func testAllowedUTIsNoneIsEmpty() {
+        let submissionTypes: [SubmissionType] = [.none]
+        let r = submissionTypes.allowedUTIs( allowedExtensions: ["png"] )
+        XCTAssertTrue(r.isEmpty)
+    }
+
+    func testAllowedUTIsAny() {
+
+        let submissionTypes: [SubmissionType] = [.online_upload]
+
+        XCTAssertEqual(submissionTypes.allowedUTIs(allowedExtensions: []), [.any])
+    }
+
+    func testAllowedUTIsAllowedExtensions() {
+        let submissionTypes: [SubmissionType] = [.online_upload]
+        let allowedExtensions = ["png", "mov", "mp3"]
+        let result = submissionTypes.allowedUTIs(allowedExtensions: allowedExtensions)
+        XCTAssertEqual(result.count, 3)
+        XCTAssertTrue(result.contains { $0.isImage })
+        XCTAssertTrue(result.contains { $0.isVideo })
+        XCTAssertTrue(result.contains { $0.isAudio })
+    }
+
+    func testAllowedUTIsAllowedExtensionsVideo() {
+        let submissionTypes: [SubmissionType] = [.online_upload]
+        let allowedExtensions = ["mov", "mp4"]
+        let result = submissionTypes.allowedUTIs(allowedExtensions: allowedExtensions)
+        XCTAssertTrue(result[0].isVideo)
+        XCTAssertTrue(result[1].isVideo)
+    }
+
+    func testAllowedUTIsMediaRecording() {
+        let submissionTypes: [SubmissionType] = [.media_recording]
+        let result = submissionTypes.allowedUTIs(allowedExtensions: [])
+        XCTAssertTrue(result.contains(.video))
+        XCTAssertTrue(result.contains(.audio))
+    }
+
+    func testAllowedUTIsText() {
+        let submissionTypes: [SubmissionType] = [.online_text_entry]
+        XCTAssertEqual(submissionTypes.allowedUTIs(allowedExtensions: []), [.text])
+    }
+
+    func testAllowedUTIsURL() {
+        let submissionTypes: [SubmissionType] = [.online_url]
+        XCTAssertEqual(submissionTypes.allowedUTIs(allowedExtensions: []), [.url])
+    }
+
+    func testAllowedUTIsMultipleSubmissionTypes() {
+        let submissionTypes: [SubmissionType] = [
+            .online_upload,
+            .online_text_entry,
+        ]
+        let allowedExtensions = ["jpeg"]
+        let result = submissionTypes.allowedUTIs(allowedExtensions: allowedExtensions)
+        XCTAssertEqual(result.count, 2)
+        XCTAssertTrue(result.contains { $0.isImage })
+        XCTAssertTrue(result.contains(.text))
     }
 }
